refactor: merge multiple createEmptySpecification helpers into one, fix broken imports
