+++ packages/api-reference/src/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from './createEmptySpecification'
 export * from './deepMerge'
 export * from './getApiClientRequest'
 export * from './getHeadingsFromMarkdown'
