+++ packages/nuxt/playwright.config.ts
@@ -1,0 +1,12 @@
+import type { ConfigOptions } from '@nuxt/test-utils/playwright'
+import { defineConfig, devices } from '@playwright/test'
+import { fileURLToPath } from 'node:url'
+
+export default defineConfig<ConfigOptions>({
+  use: {
+    nuxt: {
+      rootDir: fileURLToPath(new URL('.', import.meta.url)),
+    },
+  },
+  // ...
+})
